Class_Task/Account.py
class Account:

    def __init__(self, accountNumber, accountName, pin):

        self.__balance = 0
        self.__accountNumber = accountNumber
        self.__accountName = accountName
        self.__pin = pin

    def deposit(self, amount):
        if amount > 0:
            self.__balance += amount

    def withdraw(self, withdrawal_amount, pin):
        self.validate_Pin(pin)
        self.validate_withdrawal(withdrawal_amount)
        self.__balance -= withdrawal_amount

    def getBalance(self):
        return self.__balance

    def checkBalance(self, pin):
        self.validate_Pin(pin)
        return self.__balance

    def validate_Pin(self, pin):
        if self.__pin != pin:
            raise ValueError("Incorrect pin")

    def update_pin(self, old_pin, new_pin):
        if self.__pin == old_pin:
            self.__pin = new_pin

    def validate_withdrawal(self, withdraw_amount):
        if withdraw_amount <= 0.0:
            raise ValueError("Invalid amount")
        if self.__balance < withdraw_amount:
            raise ValueError("Insufficient Balance")

    def get_account_number(self):
        return self.__accountNumber

    def get_account_name(self):
        return self.__accountName
